// frontend/src/types/generated/api/TestSetTemplateResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TestCase } from "../models/TestCase";

export type TestSetTemplateResponse = { id: string, workspace_id: string, name: string, description: string | null, cases: Array<TestCase>, created_at: number, updated_at: number, };